options pull no occurs on server side. Added more sectors and increase pagination of sectors table to 20
