+++ client/src/components/index.js
@@ -4,4 +4,4 @@ export * from './Form'
 export * from './WatchTable'
 export * from './Logo'
 export * from './SectorTable'
-export * from './optionsTable'
+export * from './OptionsTable'
