--- client/src/model/sector.js
@@ -1,67 +1,0 @@
-let testsData = [
-  {
-    id: 1,
-    Company: 'Nasdaq 100',
-    Stock: '$QQQ',
-    Health: 90,
-    userPreference: true
-  },
-  {
-    id: 2,
-    Company: 'Healthcare',
-    Stock: '$XLV',
-    Health: 89,
-    userPreference: true
-  },
-  {
-    id: 3,
-    Company: 'SaaS',
-    Stock: '$IGV',
-    Health: 84,
-    userPreference: true
-  },
-  {
-    id: 4,
-    Company: 'Speculative Biotech',
-    Stock: '$XBI',
-    Health: 68,
-    userPreference: true
-  },
-  {
-    id: 5,
-    Company: 'Energy',
-    Stock: '$XLE',
-    Health: 24,
-    userPreference: false
-  },
-  {
-    id: 6,
-    Company: 'Small Caps',
-    Stock: '$IWM',
-    Health: 52,
-    userPreference: false
-  },
-  {
-    id: 7,
-    Company: 'China 25',
-    Stock: '$FXI',
-    Health: 42,
-    userPreference: false
-  },
-  {
-    id: 8,
-    Company: 'Semiconductors',
-    Stock: '$SMH',
-    Health: 96,
-    userPreference: false
-  },
-  {
-    id: 9,
-    Company: 'Microcaps',
-    Stock: '$IWC',
-    Health: 50,
-    userPreference: false
-  }
-]
-
-export { testsData }
