+++ client/src/model/testSector.json
@@ -36,7 +36,7 @@
   },
   {
     "id": 6,
-    "Company": "Small Caps",
+    "Company": "Small/Mid Caps",
     "Stock": "IWM",
     "Health": 52,
     "userPreference": false
@@ -59,6 +59,83 @@
     "id": 9,
     "Company": "Microcaps",
     "Stock": "IWC",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 10,
+    "Company": "SP-500",
+    "Stock": "SPY",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 11,
+    "Company": "Dow Jones Industrials",
+    "Stock": "DIA",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 12,
+    "Company": "Motley Fool Small Cap Growth",
+    "Stock": "MFMS",
+    "Health": 12,
+    "userPreference": false
+  },
+  {
+    "id": 13,
+    "Company": "Financials",
+    "Stock": "XLF",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 14,
+    "Company": "Technology",
+    "Stock": "XLK",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 15,
+    "Company": "Nasdaq Biotech Index",
+    "Stock": "IBB",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 16,
+    "Company": "Emerging Markets",
+    "Stock": "EEM",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 17,
+    "Company": "Cyber Security",
+    "Stock": "HACK",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 18,
+    "Company": "Mobile Payments ETF",
+    "Stock": "IPAY",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 19,
+    "Company": "Pet Care ETF",
+    "Stock": "PAWZ",
+    "Health": 50,
+    "userPreference": false
+  },
+  {
+    "id": 20,
+    "Company": "Cloud Computing ETF",
+    "Stock": "CLOU",
     "Health": 50,
     "userPreference": false
   }
