+++ routes/apiRoutes.js
@@ -1,6 +1,8 @@
 const router = require("express").Router();
 const db = require("../models");
 const jwt = require('jsonwebtoken');
+var cheerio = require("cheerio");
+const axios = require("axios");
 
 router.route("/api/user").post((req, res, next) => {
     db.User.create(req.body)
@@ -182,8 +184,6 @@ router.route("/api/updateSectors").put((req, res) => {
     sectorData = {
         indexName: req.body.symbol
     }
-    console.log("REQ BODY")
-    console.log(req.body)
     db.Sector.findOneAndUpdate(sectorData, {
         "priceData": req.body.priceData,
         "macdData": req.body.macdData,
@@ -193,11 +193,70 @@ router.route("/api/updateSectors").put((req, res) => {
     }).then(function (res) {
         console.log("RESPONSE")
         console.log(res)
-    })
-        .catch(function (err) {
+    }).catch(function (err) {
             // If an error occurred, send it to the client
             res.json(err);
+    })
+})
+
+router.route("/api/optionsPull/:stockTicker").get((req, res) => {
+    let ticker = req.params.stockTicker
+    let result = {
+        dates: [],
+        optionType: [],
+        strikePrice: [],
+        contractPrice: [],
+        todayChange: [],
+        volume: [],
+        openInt: [],
+        volOI: []
+    };
+    const resultRows = []
+
+    console.log("SCRAPING NASDAQ ACTIVE OPTIONS...")
+
+    async function optPull () {
+        const optionsData = axios.get(`https://old.nasdaq.com/symbol/${ticker}/option-chain/most-active`).then(response => {
+            var $ = cheerio.load(response.data);
+            $("tbody tr").each(function (i, element) {
+                result.dates.push($("td.most-active-col1").find("a").text());
+                result.optionType.push($(element).find("td.most-active-col1").next("td").text());
+                result.strikePrice.push($(element).find("td.most-active-col1").next("td").next("td").next("td").text());
+                result.contractPrice.push($(element).find("td.most-active-col1").next("td").next("td").next("td").next("td").text());
+                result.todayChange.push($(element).find("td.most-active-col1").next("td").next("td").next("td").next("td").next("td").text());
+                result.volume.push($(element).find("td.most-active-col1").next("td").next("td").next("td").next("td").next("td").next("td").next("td").text());
+                result.openInt.push($(element).find("td.most-active-col1").next("td").next("td").next("td").next("td").next("td").next("td").next("td").next("td").text());
+            });
+            function chunkString(str, length) {
+                let string = str.match(new RegExp('.{1,' + length + '}', 'g'));
+                return string;
+            }
+            result.dates = chunkString(result.dates[0], 12);
+            result.optionType = result.optionType.filter(el => { return el != ""; });
+            result.strikePrice = result.strikePrice.filter(el => { return el != ""; });
+            result.contractPrice = result.contractPrice.filter(el => { return el != ""; });
+            result.todayChange = result.todayChange.filter(el => { return el != ""; });
+            result.volume = result.volume.filter(el => { return el != ""; });
+            result.openInt = result.openInt.filter(el => { return el != ""; });
+    
+            for (let i = 0; i < result.optionType.length; i++){
+                let volOI = (result.volume[i]/result.openInt[i]*100).toFixed(2)
+                result.volOI.push(volOI + "%")
+            }
+    
+            for (let x = 0; x < result.optionType.length; x++){
+                let newResultRow = []
+                newResultRow.push(result.dates[x], result.optionType[x], result.strikePrice[x], result.contractPrice[x], result.todayChange[x], result.volume[x], result.openInt[x], result.volOI[x])
+                resultRows.push(newResultRow)
+            }
+            return resultRows
+        }).then(data => {
+            res.json(data)
+        }).catch(function (err){
+            console.log(err)
         })
+    }
+    optPull()
 })
 
 module.exports = router;
